Adding implementation for displaying word in second activty upon click and changing type for ID from Long to Int

+++ app/src/main/java/com/ranit/android/lexicon/model/Model.kt
@@ -14,7 +14,7 @@ interface Model {
     fun removeWord(id: Long) : Boolean
 
     @Throws(Exception::class)
-    fun getWord(id: Long) : Word?
+    fun getWord(id: Int) : Word?
 
     @Throws(Exception::class)
     fun getAllWords() : List<Word>?
